requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3
aiofiles==23.2.1

# Image Processing & Face Analysis
opencv-python-headless==4.9.0.80
mediapipe==0.10.9
numpy==1.26.3
Pillow==10.2.0

# Face Swap (InsightFace)
insightface==0.7.3
onnxruntime==1.16.3

# Internationalization
babel==2.14.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0

# Production
gunicorn==21.2.0
